requirements.txt
pytorch
pymatgen